api/github/db: move inmemory implementations to the db
Created with Sturdy
